+++ src/query-builder/DeleteQueryBuilder.ts
@@ -10,6 +10,7 @@ import {Brackets} from "./Brackets";
 import {DeleteResult} from "./result/DeleteResult";
 import {ReturningStatementNotSupportedError} from "../error/ReturningStatementNotSupportedError";
 import {SqljsDriver} from "../driver/sqljs/SqljsDriver";
+import {MysqlDriver} from "../driver/mysql/MysqlDriver";
 import {BroadcasterResult} from "../subscriber/BroadcasterResult";
 import {EntitySchema} from "../index";
 
@@ -66,25 +67,16 @@ export class DeleteQueryBuilder<Entity> extends QueryBuilder<Entity> implements 
             const deleteResult = new DeleteResult();
             const result = await queryRunner.query(sql, parameters);
 
-            switch (queryRunner.connection.name) {
-                case "mysql":
-                case "mariadb": {
-                    deleteResult.raw = result;
-                    deleteResult.affected = result.affectedRows;
-                    break;
-                }
-                case "mssql":
-                case "postgres": {
-                    deleteResult.raw = result[0] ? result[0] : null;
-                    // don't return 0 because it could confuse. null means that we did not receive this value
-                    deleteResult.affected = typeof result[1] === "number" ? result[1] : null;
-                    break;
-                }
-                // sqlite & sqljs doesn't return anything
-                case "sqlite":
-                case "sqljs":
-                default:
-                    deleteResult.raw = result;
+            const driver = queryRunner.connection.driver;
+            if (driver instanceof MysqlDriver) {
+                deleteResult.raw = result;
+                deleteResult.affected = result.affectedRows;
+            } else if (driver instanceof SqlServerDriver || driver instanceof PostgresDriver) {
+                deleteResult.raw = result[0] ? result[0] : null;
+                // don't return 0 because it could confuse. null means that we did not receive this value
+                deleteResult.affected = typeof result[1] === "number" ? result[1] : null;
+            } else {
+                deleteResult.raw = result;
             }
 
             // call after deletion methods in listeners and subscribers